third_party/WebKit/LayoutTests/http/tests/inspector/network/from-disk-cache-timing-expected.txt

CONSOLE MESSAGE: line 1: Done.
CONSOLE MESSAGE: line 1: Done.
Tests requests loaded from disk cache have correct timing

URL:http://127.0.0.1:8000/inspector/network/resources/cached-script.php
from memory cache: false
from disk cache: false
has timing: true

URL:http://127.0.0.1:8000/inspector/network/resources/cached-script.php
from memory cache: false
from disk cache: true
has timing: true

Second request starts later than first: true
